requirement.txt
streamlit==1.15.1
pandas==1.5.2
numpy==1.23.5
scikit-learn==1.1.3
pillow
keras
tensorflow==2.11.0
opencv-python==4.6.0.66